+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.4.0
+
+Adds support for data shape inheritance. [stefan-lacatus](https://github.com/stefan-lacatus))
+
+Resolves an issue where using the transformer with `ts.transform` would throw an error when attempting to resolve constant expressions.
+
+The transformer will no longer emit members with the `declare` modifier.  [stefan-lacatus](https://github.com/stefan-lacatus))
+
+Resolves an issue where the `__values` helper was in some cases not inlined, preventing the use of transpiled es6 features.
+
+Resolves an issue where having global functions enabled would also cause the transformer to inline functions declared in global code.
+
 # 1.3.1
 
 Resolves an issue where inline SQL statements would compile into code with syntax errors.